fix refresh token controller

+++ src/features/auth/controllers/logoutController.ts
@@ -1,20 +1,24 @@
 import { Request, Response } from "express"
 import { authService } from "../services/authService";
 import { securityRepository } from "../../security/repositories/securityRepository";
+import { jwtService } from "../../../application/jwtService";
 
 export const logoutController = async (req: Request, res: Response) => {
     const refreshToken = req.cookies.refreshToken
 
     if (!refreshToken) {
-        return res.status(401).send()
+        res.status(401).send()
+        return
+    }
+    if(jwtService.verifyRefreshToken(refreshToken)) {
+        res.status(401).send()
+        return
     }
     const deviceId = req.cookies.deviceId
     await securityRepository.deleteDeviceById(deviceId)
+
     
-    // if(!await authService.verifyRefreshToken(refreshToken)) {
-    //     return res.status(401).send()
-    // }
-    
+
     // await authService.addToExpiredTokens(req.cookies.refreshToken)
 
     res.clearCookie('refreshToken').status(204).send();
